Moved TrelloWebHookSourcesConfigFileReader to Trello project
With referenced package
