+++ ReadingList.Carter/CompositionRoot.cs
@@ -5,6 +5,7 @@ using ReadingList.Carter.Logging;
 using ReadingList.Trello.Models;
 using ReadingList.Carter.Trello;
 using Readinglist.Notes.Logic.Services;
+using ReadingList.Trello.Services;
 
 namespace ReadingList.Carter
 {
