+++ ReadingList.Carter/ReadingList.Carter.csproj
@@ -12,7 +12,6 @@
     <PackageReference Include="Carter" Version="6.0.0" />
     <PackageReference Include="LightInject.Microsoft.Hosting" Version="1.3.0" />
     <PackageReference Include="Swashbuckle.AspNetCore.SwaggerUI" Version="6.3.0" />
-    <PackageReference Include="System.Configuration.ConfigurationManager" Version="6.0.0" />
     <PackageReference Include="Serilog" Version="2.10.0" />
     <PackageReference Include="Serilog.Sinks.Console" Version="4.0.1" />
   </ItemGroup>
