+++ ReadingList.Trello/ReadingList.Trello.csproj
@@ -16,6 +16,7 @@
   <ItemGroup>
     <PackageReference Include="Manatee.Trello" Version="4.4.0" />
     <PackageReference Include="Manatee.Trello.WebApi" Version="2.0.3" />
+    <PackageReference Include="Microsoft.Extensions.Configuration.Abstractions" Version="6.0.0" />
     <PackageReference Include="Newtonsoft.Json" Version="13.0.1" />
     <PackageReference Include="Microsoft.AspNet.WebApi.Client" Version="5.2.7" />
     <PackageReference Include="LightInject" Version="6.4.1" />
